Don't panic on corrupt output in `check_sizeof`.
This gets rid of the last unnecessary panic in `check_sizeof`. There is
currently no test for the error path here because it takes a bit of work
to trigger. The most likely case where this might occur "in the wild"
would be if the user is dealing with non-UTF-8 strings, and doesn't
translate them before `probe-c-api` attempts to interpret one of them.